Séparation de la définition des layers
