+++ src/components/Map/components/layers/communesFortPotentielPourCreationReseauxChaleur.ts
@@ -1,0 +1,48 @@
+import { type Interval } from '@/utils/interval';
+
+import { type MapSourceLayersSpecification } from './common';
+
+export const communesFortPotentielPourCreationReseauxChaleurLayerColor = '#FF8329';
+export const communesFortPotentielPourCreationReseauxChaleurLayerOpacity = 0.7;
+
+export const communesFortPotentielPourCreationReseauxChaleurInterval: Interval = [0, 100_000];
+
+export const communesFortPotentielPourCreationReseauxChaleurLayersSpec = [
+  {
+    sourceId: 'communesFortPotentielPourCreationReseauxChaleur',
+    source: {
+      type: 'vector',
+      tiles: ['/api/map/communesFortPotentielPourCreationReseauxChaleur/{z}/{x}/{y}'],
+      minzoom: 5,
+      maxzoom: 6,
+    },
+    layers: [
+      {
+        id: 'communesFortPotentielPourCreationReseauxChaleur',
+        type: 'circle',
+        layout: {
+          'circle-sort-key': ['-', ['get', 'population']],
+        },
+        paint: {
+          'circle-color': communesFortPotentielPourCreationReseauxChaleurLayerColor,
+          'circle-radius': [
+            'interpolate',
+            ['linear'],
+            ['+', ['get', 'zones_fort_potentiel_chauf_mwh'], ['get', 'zones_fort_potentiel_ecs_mwh']],
+            0,
+            4,
+            160_000, // ~ max value
+            20,
+          ],
+          'circle-opacity': communesFortPotentielPourCreationReseauxChaleurLayerOpacity,
+        },
+        filter: (config) => [
+          'all',
+          ['>=', ['get', 'population'], config.communesFortPotentielPourCreationReseauxChaleur.population[0]],
+          ['<=', ['get', 'population'], config.communesFortPotentielPourCreationReseauxChaleur.population[1]],
+        ],
+        isVisible: (config) => config.communesFortPotentielPourCreationReseauxChaleur.show,
+      },
+    ],
+  },
+] as const satisfies ReadonlyArray<MapSourceLayersSpecification>;
