--- src/components/Map/map-styles.ts
@@ -1,19 +1,0 @@
-export const communesFortPotentielPourCreationReseauxChaleurLayerColor = '#FF8329';
-export const communesFortPotentielPourCreationReseauxChaleurLayerOpacity = 0.7;
-
-export const enrrMobilisablesGeothermieProfondeLayerColor = '#65a3d6';
-export const enrrMobilisablesGeothermieProfondeLayerOpacity = 0.5;
-
-export const enrrMobilisablesThalassothermieLayerColor = '#4c64c9';
-export const enrrMobilisablesThalassothermieLayerOpacity = 0.6;
-
-export const installationsGeothermieProfondeLayerColor = '#8400a8';
-export const installationsGeothermieProfondeLayerOpacity = 0.8;
-
-export const installationsGeothermieSurfaceEchangeursFermesRealiseeColor = '#b96210';
-export const installationsGeothermieSurfaceEchangeursFermesDeclareeColor = '#c6c614';
-export const installationsGeothermieSurfaceEchangeursFermesOpacity = 0.8;
-
-export const installationsGeothermieSurfaceEchangeursOuvertsRealiseeColor = '#3217f0';
-export const installationsGeothermieSurfaceEchangeursOuvertsDeclareeColor = '#17f0f0';
-export const installationsGeothermieSurfaceEchangeursOuvertsOpacity = 0.8;
