+++ src/components/NetworksList/NetworksList.tsx
@@ -5,7 +5,7 @@ import { useCallback, useEffect, useMemo, useState } from 'react';
 import styled from 'styled-components';
 import XLSX from 'xlsx';
 
-import { reseauxDeChaleurFilters } from '@/components/Map/map-layers';
+import { reseauxDeChaleurFilters } from '@/components/Map/components/layers/filters';
 import ReseauxDeChaleurFilters, { type ReseauxDeChaleurFiltersProps } from '@/components/ReseauxDeChaleurFilters';
 import Box from '@/components/ui/Box';
 import Drawer from '@/components/ui/Drawer';
