--- src/services/Map/businessRules/buildings.ts
@@ -1,20 +1,0 @@
-import { type ExpressionInputType } from 'maplibre-gl';
-
-export const themeDefBuildings = {
-  colors: {
-    a: { color: '#0D8A61' },
-    b: { color: '#42A548' },
-    c: { color: '#6CB36E' },
-    d: { color: '#EBDE2D' },
-    e: { color: '#E0A736' },
-    f: { color: '#E66E31' },
-    g: { color: '#C5171D' },
-    n: { color: '#999999' },
-    unknow: { color: '#999999' },
-  },
-  opacity: 0.65,
-};
-
-export const arrColorFromDefBuildingsDpeEnergy = Object.entries(themeDefBuildings.colors).flatMap(
-  ([dpeCode, dpeStyleDef]: [string, any]) => [dpeCode, dpeStyleDef.color]
-) as [ExpressionInputType, ExpressionInputType, ...ExpressionInputType[]];
