--- src/services/Map/businessRules/energy.ts
@@ -1,15 +1,0 @@
-export const localTypeEnergy: Record<string, string> = {
-  fuelOil: 'Bâtiments chauffés au fioul collectif',
-  gas: 'Bâtiments chauffés au gaz collectif',
-  wood: 'Bois',
-  electric: 'Electrique',
-  unknow: 'Autre',
-};
-
-export const themeDefEnergy = {
-  fuelOil: { color: '#F07300' },
-  gas: { color: '#9181F4' },
-  wood: { color: '#955e15' },
-  electric: { color: '#4cd362' },
-  unknow: { color: '#818181' },
-};
