--- src/services/Map/businessRules/enrrMobilisables.ts
@@ -1,9 +1,0 @@
-export const themeDefSolaireThermiqueFriches = {
-  color: '#dc958e',
-  opacity: 0.7,
-} as const;
-
-export const themeDefSolaireThermiqueParkings = {
-  color: '#d0643e',
-  opacity: 0.7,
-} as const;
