--- src/services/Map/businessRules/gasUsage.ts
@@ -1,13 +1,0 @@
-export const localTypeGas = {
-  T: 'Tertiaire',
-  R: 'Logements (tous types)',
-  I: 'Industrie',
-  unknow: 'Inconnu',
-};
-
-export const themeDefTypeGas = {
-  T: { color: '#0032E5' },
-  R: { color: '#00B8F0' },
-  I: { color: '#00efaf' },
-  unknow: { color: '#818181' },
-};
