--- src/services/Map/businessRules/heatNetwork.ts
@@ -1,7 +1,0 @@
-export const themeDefHeatNetwork = {
-  outline: { color: '#7CC558' },
-  futur: { color: '#DA5DD5', opacity: 0.47 },
-  classed: { color: '#079067' },
-  cold: { color: '#0094FF' },
-  unknow: { color: '#818181' },
-};
