--- src/services/Map/businessRules/index.ts
@@ -1,6 +1,0 @@
-export { themeDefBuildings } from './buildings';
-export { themeDefDemands } from './demands';
-export { localTypeEnergy, themeDefEnergy } from './energy';
-export { localTypeGas, themeDefTypeGas } from './gasUsage';
-export { themeDefHeatNetwork } from './heatNetwork';
-export { themeDefZoneDP } from './zoneDP';
