--- src/services/Map/businessRules/zoneDP.ts
@@ -1,3 +1,0 @@
-export const themeDefZoneDP = {
-  fill: { color: '#f0bb00', opacity: 0.47 },
-};
