--- src/services/Map/businessRules/zonePotentielChaud.ts
@@ -1,6 +1,0 @@
-export const themeDefZonePotentielChaud = {
-  fill: { color: '#b0cc4e', opacity: 0.3 },
-} as const;
-export const themeDefZonePotentielFortChaud = {
-  fill: { color: '#448d60', opacity: 0.3 },
-} as const;
